tfbuilder: refactor input fairmq backend
